chore: make FileEntry at table mode to work
